Renamed services to python-case naming and imports adjusted
